\name{measureA}
\alias{measureA}
\title{
Calculate the Vargha-Delaney A measure between two populations
}
\description{
Calculate the Vargha-Delaney A measure between two populations
}
\usage{
measureA(a, b)
}

\arguments{
  \item{a}{
vector: the first population
}
  \item{b}{
vector: the second population
}
}
\details{
}
\value{
real: The A measure
}
\references{
}
\author{
Stephen W. Thomas
}
\note{
}

\seealso{
}
\examples{
}
